Retroactively uploading files from 2015
This was a project I led in 2015 for a Drexel class. After six months we had a working prototype of a mobile application that displayed short text posts. The text was displayed on post-it notes that appeared relative to the user's physical location (i.e. posts made physically North of you would appear above you on screen). Users could "upvote" and "downvote" other posts by placing adhesives like tacks and tape.